move components into libs folder, refactor
